MHP-1639: Add public survey form
- Cleanup gray color naming
- Move color class generation to new file to prevent duplicate class creation
