+++ app/assets/javascripts/angular/components/organizationContactImport/organizationContactImport.html
@@ -4,7 +4,7 @@
             <div class="w-100 items-end">
                 <h1 class="f2 normal blue">Import Contacts</h1>
                 <div class="progress flex items-center justify-center center">
-                    <div class="progress-step br-100 ba bw3 b--grey-light"
+                    <div class="progress-step br-100 ba bw3 b--gray-light"
                          ng-repeat="step in [1, 2, 3]" ng-class="{'bg-blue white': step === $ctrl.activeStep}">
                         {{step}}
                     </div>
