+++ app/assets/javascripts/angular/components/organizationContactImport/organizationContactImportStep2.html
@@ -1,11 +1,11 @@
-<div class="import-contact-form mt5 bb b--grey-medium pb3 mb4">
+<div class="import-contact-form mt5 bb b--gray-medium pb3 mb4">
     <h2 class="normal f3 fw6 blue">{{ 'contact_import:step_2.title' | t }}</h2>
-    <p class="lh-copy grey-darker f4 w-90">
+    <p class="lh-copy gray-darker f4 w-90">
         {{ 'contact_import:step_2.instructions' | t }}
     </p>
-    <table class="collapse w-100 f6 grey-darker">
+    <table class="collapse w-100 f6 gray-darker">
         <thead>
-        <tr class="bg-grey-light ttu tracked bold">
+        <tr class="bg-gray-light ttu tracked bold">
             <th class="ph3 pv2">{{ 'contact_import:step_2.csv_column' | t }}</th>
             <th class="ph3 pv2 flex items-center">
                 <a href="" ng-click="$ctrl.previousPreviewRow()">
@@ -21,11 +21,11 @@
         </thead>
         <tbody>
         <tr ng-repeat="(index, value) in $ctrl.csvData[0] track by $index">
-            <td class="ph3 pb1 pt3 grey-darkest">{{$ctrl.csvData[0][index]}}</td>
-            <td class="ph3 pb1 pt3 grey-darkest">{{$ctrl.csvData[$ctrl.currentPreviewRow][index]}}</td>
-            <td class="ph3 pb1 pt3 grey-darkest">
-                <div class="dropdown dropdown-select w-100 mr2 icon-grey-darker" uib-dropdown>
-                    <a class="f6 link db pa3 input-reset bg-white ba grey-darker w-100 br2 b--grey-dark dim" href=""
+            <td class="ph3 pb1 pt3 gray-darkest">{{$ctrl.csvData[0][index]}}</td>
+            <td class="ph3 pb1 pt3 gray-darkest">{{$ctrl.csvData[$ctrl.currentPreviewRow][index]}}</td>
+            <td class="ph3 pb1 pt3 gray-darkest">
+                <div class="dropdown dropdown-select w-100 mr2 icon-gray-darker" uib-dropdown>
+                    <a class="f6 link db pa3 input-reset bg-white ba gray-darker w-100 br2 b--gray-dark dim" href=""
                        uib-dropdown-toggle>
                         {{$ctrl.columnMap[index] ?
                         ($ctrl.columnMap[index].column_title || $ctrl.columnMap[index].label) :
@@ -58,7 +58,7 @@
 
 <div class="buttons">
     <button ng-click="$ctrl.previous()"
-            class="f5 link dim ph5 w-33 pv3 mb2 dib white fl bold br-pill ttu tracked bg-grey-darker">
+            class="f5 link dim ph5 w-33 pv3 mb2 dib white fl bold br-pill ttu tracked bg-gray-darker">
         {{ 'contact_import:back' | t }}
     </button>
     <button class="f5 link dim ph5 pv3 w-33 mb2 dib white fr bold br-pill ttu tracked bg-blue"
