+++ app/assets/javascripts/angular/components/organizationContactImport/organizationContactImportStep4.html
@@ -4,7 +4,7 @@
         <h1 class="f2 normal blue mb3 mt3">{{ 'contact_import:step_4.title' | t }}</h1>
     </header>
     <div class="mt0 tc pb3 mb4">
-        <p class="lh-copy grey-darker f5 tc mb4">
+        <p class="lh-copy gray-darker f5 tc mb4">
             {{ 'contact_import:step_4.instructions' | t }}
         </p>
         <p class="tc">
