+++ app/assets/javascripts/angular/components/publicSurvey/publicSurvey.component.js
@@ -1,0 +1,8 @@
+import template from './publicSurvey.html';
+
+angular.module('missionhubApp').component('publicSurvey', {
+    bindings: {
+        survey: '<',
+    },
+    template,
+});
