+++ app/assets/javascripts/angular/components/surveyOverviewKeywordTab/keywordRequestModal.html
@@ -4,7 +4,7 @@
             <img class="dib" ng-src="{{$ctrl.clockIcon}}" style="width:60px; height:60px;">
             <h2 class="f3 normal blue mb2 mt2">{{'surveys:keyword.requested' | t}}</h2>
         </div>
-        <p class="lh-copy measure center f5 grey-darker mb0 mt0">
+        <p class="lh-copy measure center f5 gray-darker mb0 mt0">
             {{'surveys:keyword.requested_message' | t}}
         </p>
     </div>
