+++ app/assets/javascripts/angular/components/surveyOverviewQuestionsTab/surveyOverviewQuestions.html
@@ -1,9 +1,9 @@
 <div class="accordion">
-    <form class="bg-grey-light br2 grey-darker" ng-repeat="question in $ctrl.surveyQuestions track by question.id"
+    <form class="bg-pivot-gray-5 br2 pivot-gray-darker" ng-repeat="question in $ctrl.surveyQuestions track by question.id"
         ng-class="{
             'accordion-expanded': $ctrl.isExpanded[question.id]
         }">
-        <h2 class="accordion-header sortable normal grey-darker mt3 bg-grey-light f5 flex br2"
+        <h2 class="accordion-header sortable normal pivot-gray-darker mt3 bg-pivot-gray-5 f5 flex br2"
             ng-class="{
                 'pa3': !$ctrl.isExpanded[question.id],
                 'ph3 pv2 items-center': $ctrl.isExpanded[question.id]
@@ -15,7 +15,7 @@
                 {{question.label}}
                 <span class="f7" ng-if="question.required">({{ 'common:general.required' | t }})</span>
             </span>
-            <input class="pa2 input-reset bg-white bw0 grey-darker w-83 ihshort br2"
+            <input class="pa2 input-reset bg-white bw0 pivot-gray-darker w-83 ihshort br2"
                    placeholder="{{ 'surveys:questions.labelPlaceholder' | t }}" type="text" ng-model="question.label"
                    ng-change="$ctrl.saveQuestion(question)"
                    ng-if="$ctrl.isExpanded[question.id]">
@@ -34,11 +34,11 @@
             <fieldset class="ba b--transparent ph0 mh0 w-91">
                 <div class="flex justify-between mb3">
                     <div class="w-48">
-                        <label class="db f6 lh-copy f5 normal mb1 grey-darker ttu tracked">
+                        <label class="db f6 lh-copy f5 normal mb1 pivot-gray-darker ttu tracked">
                             {{ 'surveys:questions.type' | t }}
                         </label>
                         <div class="dropdown dropdown-select w-100 mr2" uib-dropdown>
-                            <a class="f5 link dim bg-white ihshort db grey-darker br2 dim" href="" uib-dropdown-toggle>
+                            <a class="f5 link dim bg-white ihshort db pivot-gray-darker br2 dim" href="" uib-dropdown-toggle>
                                 {{$ctrl.getQuestionType(question.kind, question.style).name}}
                             </a>
                             <ul class="bg-white f5" uib-dropdown-menu>
@@ -52,17 +52,17 @@
                         </div>
                     </div>
                     <div class="w-48">
-                        <label class="db f5 lh-copy f6 normal mb1 grey-darker ttu tracked">
+                        <label class="db f5 lh-copy f6 normal mb1 pivot-gray-darker ttu tracked">
                             {{ 'surveys:questions.columnTitle' | t }}
                         </label>
-                        <input class="pa2 input-reset bg-white bw0 grey-darker w-100 ihshort br2"
+                        <input class="pa2 input-reset bg-white bw0 pivot-gray-darker w-100 ihshort br2"
                                ng-change="$ctrl.saveQuestion(question)"
                                ng-model="question.column_title" type="text">
                     </div>
                 </div>
                 <div class="flex justify-between">
                     <div class="w-100">
-                        <!--<p class="db f6 lh-copy f6 normal mb1 grey-darker ttu fw6 tracked mv3">-->
+                        <!--<p class="db f6 lh-copy f6 normal mb1 pivot-gray-darker ttu fw6 tracked mv3">-->
                             <!--Answers to initiate assignment/notifications-->
                         <!--</p>-->
                         <div class="assignments pv2"
@@ -77,7 +77,7 @@
                                     </a>
                                 </div>
                                 <div class="f-grow">
-                                    <input class="pa2 input-reset bg-white bw0 grey-darker w-100 ihshort br2" type="text"
+                                    <input class="pa2 input-reset bg-white bw0 pivot-gray-darker w-100 ihshort br2" type="text"
                                            ng-change="$ctrl.saveQuestionContent(question, questionAnswers)"
                                            ng-model="questionAnswers[$index]">
                                 </div>
@@ -99,7 +99,7 @@
                     <!--<div class="w-48" ng-if="questionAnswers.length" ng-init="answerTab = 0">-->
                         <!--<div class="tabs flex justify-between">-->
                             <!--<div class="w-49 relative">-->
-                                <!--<a class="tab bl br bt pv3 ph2 link grey-darker br2 br&#45;&#45;top db b&#45;&#45;grey-dark"-->
+                                <!--<a class="tab bl br bt pv3 ph2 link pivot-gray-darker br2 br&#45;&#45;top db b&#45;&#45;grey-dark"-->
                                    <!--href="" ng-click="answerTab = 0" ng-class="{'active': answerTab === 0}">-->
                                     <!--Auto Assign-->
                                 <!--</a>-->
@@ -108,7 +108,7 @@
                                 <!--</a>-->
                             <!--</div>-->
                             <!--<div class="w-49 relative">-->
-                                <!--<a class="tab bl br bt pv3 ph2 link grey-darker br2 br&#45;&#45;top db b&#45;&#45;grey-dark"-->
+                                <!--<a class="tab bl br bt pv3 ph2 link pivot-gray-darker br2 br&#45;&#45;top db b&#45;&#45;grey-dark"-->
                                    <!--href="" ng-click="answerTab = 1" ng-class="{'active': answerTab === 1}">-->
                                     <!--Auto Notify-->
                                 <!--</a>-->
@@ -138,11 +138,11 @@
                                 <!--</a>-->
                             <!--</div>-->
                             <!--<div class="flex items-center mv-ih">-->
-                                <!--<label class="db f6 lh-copy f5 normal grey-darker mr2">-->
+                                <!--<label class="db f6 lh-copy f5 normal pivot-gray-darker mr2">-->
                                     <!--Send notifications via:-->
                                 <!--</label>-->
                                 <!--<div class="dropdown dropdown-select mr2" uib-dropdown>-->
-                                    <!--<a class="f6 link dim bg-white ihshort db grey-darker br2 dim pr5" href=""-->
+                                    <!--<a class="f6 link dim bg-white ihshort db pivot-gray-darker br2 dim pr5" href=""-->
                                        <!--uib-dropdown-toggle>-->
                                         <!--{{question.notify_via}}-->
                                     <!--</a>-->
@@ -165,7 +165,7 @@
 </div>
 
 <div class="new-questions flex items-start mb2 mt3">
-    <!--<div class="dropdown w-50 mr2" uib-dropdown>-->
+    <!--<div class="dropdown-large dropdown w-50 mr2" uib-dropdown>-->
         <!--<a class="f6 link dim ba ph3 pv3 db blue br2 dim ttu tracked" href="" uib-dropdown-toggle>-->
             <!--{{ 'surveys:questions.addPredefinedQuestion' | t }}-->
         <!--</a>-->
@@ -175,7 +175,7 @@
             <!--</li>-->
         <!--</ul>-->
     <!--</div>-->
-    <div class="dropdown w-50 ml2" uib-dropdown>
+    <div class="dropdown-large dropdown w-50 ml2" uib-dropdown>
         <a class="f6 link dim ba ph3 pv3 db blue br2 dim ttu tracked" href="" uib-dropdown-toggle>
             {{ 'surveys:questions.addNewQuestion' | t }}
         </a>
