+++ app/assets/javascripts/angular/components/surveyOverviewSettingsTab/surveyOverviewSettings.html
@@ -1,29 +1,29 @@
-<div class="bg-grey-light">
+<div class="bg-gray-light">
     <div class="panel-body pt10 pb4">
         <form class="measure center">
             <fieldset class="ba b--transparent ph0 mh0" id="sign_up">
-                <div class="mt4 bb pv4 b--grey-medium tl">
+                <div class="mt4 bb pv4 b--gray-medium tl">
                     <label class="db fw6 lh-copy f5 bold mb1 blue ttu tracked" for="surveyName">
                         {{'surveyTab:surveyName' | t}}
                     </label>
-                    <input class="pa2 input-reset bg-white bw0 grey-darker w-100 ihshort" id="surveyName"
+                    <input class="pa2 input-reset bg-white bw0 gray-darker w-100 ihshort" id="surveyName"
                            placeholder="{{'surveyTab:surveyName' | t}}" type="text" ng-model="$ctrl.surveyEdit.title"
                            ng-change="$ctrl.saveSurvey()">
                 </div>
-                <div class="mt4 bb pb4 b--grey-medium tl">
+                <div class="mt4 bb pb4 b--gray-medium tl">
                     <label class="db fw6 lh-copy f5 bold mb1 blue ttu tracked" for="welcomeMessage">
                         {{'surveys:settings.welcome_message' | t}}
                     </label>
-                    <textarea class="pa2 input-reset bg-white bw0 grey-darker w-100" id="welcomeMessage"
+                    <textarea class="pa2 input-reset bg-white bw0 gray-darker w-100" id="welcomeMessage"
                               name="welcomeMessage" placeholder="{{'surveys:settings:welcome_message' | t}}"
                               ng-model="$ctrl.surveyEdit.login_paragraph"
                               ng-change="$ctrl.saveSurvey()"></textarea>
                 </div>
-                <div class="mt4 bb pb4 b--grey-medium tl">
+                <div class="mt4 bb pb4 b--gray-medium tl">
                     <label class="db fw6 lh-copy f5 bold mb1 blue ttu tracked" for="successMessage">
                         {{'surveys:settings.success_message' | t}}
                     </label>
-                    <textarea class="pa2 input-reset bg-white bw0 grey-darker w-100" id="successMessage"
+                    <textarea class="pa2 input-reset bg-white bw0 gray-darker w-100" id="successMessage"
                               placeholder="{{'surveys:settings.success_message' | t}}"
                               ng-model="$ctrl.surveyEdit.post_survey_message"
                               ng-change="$ctrl.saveSurvey()"></textarea>
@@ -31,7 +31,7 @@
                 <div class="mt4 tl">
                     <label class="db fw6 lh-copy f5 bold mb1 blue ttu tracked">
                         {{'surveys:settings.image' | t}}
-                        <span class="f6 grey-dark ttn normal">(480 x 240 .jpg, .png, .gif)</span>
+                        <span class="f6 gray-dark ttn normal">(480 x 240 .jpg, .png, .gif)</span>
                     </label>
                     <div class="image-box bg-white mb3 br2">
                         <img ng-src="{{$ctrl.surveyEdit.logo_url}}" ng-if="$ctrl.surveyEdit.logo_url">
