+++ netlify.toml
@@ -1,4 +1,9 @@
 [[redirects]]
+  from = "/s/*"
+  to = "/index.html"
+  status = 200
+
+[[redirects]]
   from = "/"
   to = "/netlify-to-rails-redirect.html"
   status = 200
